requirements.txt
fastapi[all]
sqlalchemy 
databases[postgresql]
dotenv
passlib
pyjwt